Move pimoroni trackball to drivers/sensors/ too
