+++ keyboards/draculad/keymaps/pimoroni/rules.mk
@@ -1,6 +1,6 @@
 # only uncomment on the side you have your trackball on
 POINTING_DEVICE_ENABLE = yes
-SRC += pimoroni_trackball.c
+SRC += drivers/sensors/pimoroni_trackball.c
 QUANTUM_LIB_SRC += i2c_master.c
 OLED_DRIVER_ENABLE = yes
 MOUSEKEY_ENABLE = no
