+++ layouts/community/ergodox/drashna/config.h
@@ -47,3 +47,4 @@
 
 #define PIMORONI_TRACKBALL_INVERT_X
 #define PIMORONI_TRACKBALL_INVERT_Y
+#define PIMORONI_TRACKBALL_CLICK
