+++ users/drashna/rules.mk
@@ -74,7 +74,7 @@ endif
 ifeq ($(strip $(PIMORONI_TRACKBALL_ENABLE)), yes)
     POINTING_DEVICE_ENABLE := yes
     OPT_DEFS += -DPIMORONI_TRACKBALL_ENABLE
-    SRC += pimoroni_trackball.c
+    SRC += drivers/sensors/pimoroni_trackball.c
     QUANTUM_LIB_SRC += i2c_master.c
 endif
